# generated/routes/logs.ts
// Auto-generated by RouteGenerator. Do not edit manually.
import { Hono } from 'hono';
import type { AppEnv } from 'schematic-pg/api/types';
import { validateJson, validateParam, validateQuery } from 'schematic-pg/api/middleware/validate';
import { notFoundResponse } from 'schematic-pg/api/middleware/errors';
import { buildReadQuery } from 'schematic-pg/api/utils/read-query';
import { parseIncludeQuery } from 'schematic-pg/api/utils/include-query';
import { omitFields } from 'schematic-pg/api/utils/omit-fields';
import { shapeResponse, shapeResponseMany } from 'schematic-pg/api/utils/response-shape';
import {
  LogCreateSchema,
  LogUpdateSchema,
  LogParamSchema,
  LogListQuerySchema,
  LogGetQuerySchema,
  LOG_LIST_QUERY_FIELDS,
  LOG_SORTABLE_FIELDS,
  LOG_INCLUDABLE_RELATIONS,
  LOG_OMIT_FIELDS,
  API_OMIT_FIELDS_BY_MODEL,
  API_RELATION_TARGETS,
} from '../schemas/validation.js';

const router = new Hono<AppEnv>();

router.get('/', validateQuery(LogListQuerySchema), async (c) => {
  const db = c.get('db');
  const query = c.req.valid('query');
  const { where, orderBy, take, skip, include } = buildReadQuery(
    query,
    LOG_LIST_QUERY_FIELDS,
    LOG_SORTABLE_FIELDS,
    LOG_INCLUDABLE_RELATIONS,
  );
  const rows = await db.log.findMany({
    where,
    orderBy,
    take,
    skip,
    include,
  });
  return c.json(shapeResponseMany(rows, 'Log', API_OMIT_FIELDS_BY_MODEL, API_RELATION_TARGETS));
});

router.get('/:id', validateParam(LogParamSchema), validateQuery(LogGetQuerySchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const query = c.req.valid('query');
  const include = query.include
    ? parseIncludeQuery(query.include, LOG_INCLUDABLE_RELATIONS)
    : undefined;
  const row = await db.log.findUnique({ id: params.id }, { include });
  if (!row) {
    return notFoundResponse(c);
  }
  return c.json(shapeResponse(row, 'Log', API_OMIT_FIELDS_BY_MODEL, API_RELATION_TARGETS));
});

router.post('/', validateJson(LogCreateSchema), async (c) => {
  const db = c.get('db');
  const body = c.req.valid('json');
  const row = await db.log.create(body);
  return c.json(omitFields(row, LOG_OMIT_FIELDS), 201);
});

router.put('/:id', validateParam(LogParamSchema), validateJson(LogUpdateSchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const body = c.req.valid('json');
  const row = await db.log.update({ where: { id: params.id }, data: body });
  return c.json(omitFields(row, LOG_OMIT_FIELDS));
});

router.delete('/:id', validateParam(LogParamSchema), async (c) => {
  const db = c.get('db');
  const params = c.req.valid('param');
  const row = await db.log.delete({ id: params.id });
  return c.json(omitFields(row, LOG_OMIT_FIELDS));
});

export default router;